[AMD] emulate descriptor reduce (#639)
