+++ lib/Dialect/Triton/Transforms/CMakeLists.txt
@@ -12,6 +12,7 @@ add_triton_library(TritonTransforms
   RewriteTensorDescriptorToPointer.cpp
   ArithTypeConversion.cpp
   FunctionTypeConversion.cpp
+  TensorDescriptorConversionUtils.cpp
 
   DEPENDS
   TritonTransformsIncGen
