+++ lib/Dialect/Triton/Transforms/TensorDescriptorConversionUtils.cpp
@@ -1,0 +1,586 @@
+#include "triton/Dialect/Triton/Transforms/TensorDescriptorConversionUtils.h"
+#include "triton/Dialect/Triton/Transforms/ArithTypeConversion.h"
+#include "triton/Dialect/Triton/Transforms/FunctionTypeConversion.h"
+#include "triton/Dialect/Triton/Transforms/Passes.h"
+
+#include "triton/Dialect/Triton/IR/Dialect.h"
+#include "triton/Dialect/Triton/IR/Types.h"
+
+#include "mlir/Dialect/SCF/IR/SCF.h"
+#include "mlir/Dialect/SCF/Transforms/Patterns.h"
+#include "mlir/IR/Attributes.h"
+#include "mlir/IR/BuiltinAttributes.h"
+#include "mlir/IR/Operation.h"
+#include "mlir/IR/ValueRange.h"
+#include "mlir/Support/LLVM.h"
+#include "llvm/ADT/ArrayRef.h"
+#include "llvm/ADT/STLExtras.h"
+#include "llvm/ADT/SmallVector.h"
+#include "llvm/ADT/SmallVectorExtras.h"
+#include "llvm/Support/LogicalResult.h"
+#include "llvm/Support/raw_ostream.h"
+#include <mlir/Dialect/Arith/IR/Arith.h>
+#include <mlir/Dialect/Func/Transforms/FuncConversions.h>
+#include <mlir/IR/Builders.h>
+#include <mlir/IR/Value.h>
+#include <mlir/Pass/Pass.h>
+#include <mlir/Transforms/DialectConversion.h>
+
+#include <iterator>
+
+namespace mlir::triton {
+
+namespace {
+using namespace mlir;
+
+// Filter out operand segment sizes from the list of attributes since
+// this attribute is operation specific and shouldn't be set arbitrarily.
+SmallVector<NamedAttribute> filterSegmentSizes(ArrayRef<NamedAttribute> attrs) {
+  SmallVector<NamedAttribute> ret;
+  llvm::copy_if(attrs, std::back_inserter(ret), [](const NamedAttribute &attr) {
+    auto attrName = attr.getName().getValue();
+    return attrName != "operandSegmentSizes";
+  });
+  return ret;
+}
+
+Descriptor unpackDescriptor(TensorDescType type, ValueRange pack) {
+  int rank = type.getBlockType().getRank();
+  assert(pack.size() == 1 + 2 * static_cast<size_t>(rank) + 2 &&
+         "Expected tensor descriptors to consist of a pointer, "
+         "followed by 'rank' shape values and 'rank' stride values, "
+         "followed by padding and TF32 rounding option values.");
+
+  Descriptor res;
+  res.base = pack[0];
+  res.shape = pack.slice(1, rank);
+  res.strides = pack.slice(1 + rank, rank);
+  res.paddingOption = pack[1 + 2 * rank];
+  res.roundF32ToTF32 = pack[2 + 2 * rank];
+  return res;
+}
+
+Value expandOffsets(OpBuilder &builder, Location loc,
+                    ArrayRef<int64_t> blockShape, Value offsets, unsigned dim) {
+  Value expandedResult = offsets;
+  for (size_t j = 0; j < blockShape.size(); ++j) {
+    if (j == dim) {
+      continue;
+    }
+    expandedResult =
+        triton::ExpandDimsOp::create(builder, loc, expandedResult, j);
+  }
+
+  return expandedResult;
+}
+
+Value getExpandedOffsetWithRange(OpBuilder &builder, Location loc,
+                                 ArrayRef<int64_t> blockShape, Value offset,
+                                 unsigned dim) {
+  // Add range
+  auto indexI32RowType =
+      RankedTensorType::get({blockShape[dim]}, builder.getI32Type());
+  auto indexRowType =
+      RankedTensorType::get({blockShape[dim]}, builder.getI64Type());
+  Value splatOffset =
+      triton::SplatOp::create(builder, loc, indexRowType, offset);
+  Value range = triton::MakeRangeOp::create(builder, loc, indexI32RowType, 0,
+                                            blockShape[dim]);
+  Value i64Range = arith::ExtSIOp::create(builder, loc, indexRowType, range);
+
+  Value offsets = arith::AddIOp::create(builder, loc, splatOffset, i64Range);
+  return expandOffsets(builder, loc, blockShape, offsets, dim);
+}
+
+Value generatePtrFromOffsetRanges(OpBuilder &builder, Location loc,
+                                  ArrayRef<int64_t> blockShape,
+                                  Descriptor &desc, ValueRange offsets) {
+  assert(blockShape.size() == desc.shape.size());
+  assert(blockShape.size() == offsets.size());
+  auto indexTensorType =
+      RankedTensorType::get(blockShape, builder.getI64Type());
+  auto ptrType = cast<triton::PointerType>(desc.base.getType());
+  auto ptrTensorType = RankedTensorType::get(blockShape, ptrType);
+
+  // Generate offsets per dimension
+  Value ptr = triton::SplatOp::create(builder, loc, ptrTensorType, desc.base);
+  for (unsigned i = 0; i < blockShape.size(); ++i) {
+    // We must splat strides into the expanded shape not a row for retaining
+    // the divisibility information given by strides
+    Value splatStride = triton::SplatOp::create(
+        builder, loc, offsets[i].getType(), desc.strides[i]);
+    Value offsetWithStride =
+        arith::MulIOp::create(builder, loc, offsets[i], splatStride);
+    Value broadcasted = triton::BroadcastOp::create(
+        builder, loc, indexTensorType, offsetWithStride);
+
+    // Add to the pointer
+    ptr =
+        triton::AddPtrOp::create(builder, loc, ptrTensorType, ptr, broadcasted);
+  }
+
+  return ptr;
+}
+
+Value generatePtr(OpBuilder &builder, Location loc,
+                  ArrayRef<int64_t> blockShape, Descriptor &desc,
+                  ValueRange offsets) {
+  assert(blockShape.size() == desc.shape.size());
+  assert(blockShape.size() == offsets.size());
+  SmallVector<Value> offsetRanges;
+  for (unsigned i = 0; i < blockShape.size(); ++i) {
+    auto offsetWithRange =
+        getExpandedOffsetWithRange(builder, loc, blockShape, offsets[i], i);
+    offsetRanges.push_back(offsetWithRange);
+  }
+
+  return generatePtrFromOffsetRanges(builder, loc, blockShape, desc,
+                                     offsetRanges);
+}
+
+Value generateMaskFromOffsetRanges(OpBuilder &builder, Location loc,
+                                   ArrayRef<int64_t> blockShape,
+                                   Descriptor &desc, ValueRange offsetRanges) {
+  assert(blockShape.size() == desc.shape.size());
+  assert(blockShape.size() == offsetRanges.size());
+
+  // Generate mask per dimension
+  auto maskTensorType = RankedTensorType::get(blockShape, builder.getI1Type());
+  Value mask;
+  for (std::size_t i = 0; i < blockShape.size(); ++i) {
+    auto offsetWithRange = offsetRanges[i];
+
+    // Compare with lower bound
+    Value lowerBound =
+        arith::ConstantIntOp::create(builder, loc, builder.getI64Type(), 0);
+    Value splatLowerBound = triton::SplatOp::create(
+        builder, loc, offsetWithRange.getType(), lowerBound);
+    Value cmpLower =
+        arith::CmpIOp::create(builder, loc, arith::CmpIPredicate::sge,
+                              offsetWithRange, splatLowerBound);
+
+    // Compare with upper bound
+    Value splatUpperBound = triton::SplatOp::create(
+        builder, loc, offsetWithRange.getType(), desc.shape[i]);
+    Value cmpUpper =
+        arith::CmpIOp::create(builder, loc, arith::CmpIPredicate::slt,
+                              offsetWithRange, splatUpperBound);
+
+    // And and broadcast
+    Value andResult = arith::AndIOp::create(builder, loc, cmpLower, cmpUpper);
+    Value broadcasted =
+        triton::BroadcastOp::create(builder, loc, maskTensorType, andResult);
+
+    // And up all results
+    if (!mask) {
+      mask = broadcasted;
+    } else {
+      mask = arith::AndIOp::create(builder, loc, mask, broadcasted);
+    }
+  }
+
+  return mask;
+}
+
+Value generateMask(OpBuilder &builder, Location loc,
+                   ArrayRef<int64_t> blockShape, Descriptor &desc,
+                   ValueRange offsets) {
+  assert(blockShape.size() == desc.shape.size());
+  assert(blockShape.size() == offsets.size());
+  SmallVector<Value> offsetRanges;
+  for (unsigned i = 0; i < blockShape.size(); ++i) {
+    auto offsetWithRange =
+        getExpandedOffsetWithRange(builder, loc, blockShape, offsets[i], i);
+    offsetRanges.push_back(offsetWithRange);
+  }
+
+  return generateMaskFromOffsetRanges(builder, loc, blockShape, desc,
+                                      offsetRanges);
+}
+
+Value generateOther(OpBuilder &builder, Location loc, Type scalarTy,
+                    ArrayRef<int64_t> blockShape,
+                    Value paddingOption = nullptr) {
+  auto blockTy = RankedTensorType::get(blockShape, scalarTy);
+  if (paddingOption && isa<FloatType>(scalarTy)) {
+    auto floatTy = cast<FloatType>(scalarTy);
+    auto nan = llvm::APFloat::getNaN(floatTy.getFloatSemantics());
+    auto nanValue = arith::ConstantOp::create(
+        builder, loc,
+        SplatElementsAttr::get(blockTy, builder.getFloatAttr(floatTy, nan)));
+    auto zeroValue = arith::ConstantOp::create(
+        builder, loc,
+        SplatElementsAttr::get(blockTy, builder.getZeroAttr(floatTy)));
+    return arith::SelectOp::create(builder, loc, paddingOption, nanValue,
+                                   zeroValue);
+  } else {
+    auto attr = builder.getZeroAttr(blockTy);
+    return arith::ConstantOp::create(builder, loc, attr);
+  }
+}
+
+Value generateOther(OpBuilder &builder, Location loc, TensorDescType descTy,
+                    Value paddingOption = nullptr) {
+  auto blockTy = descTy.getSignlessBlockType();
+  return generateOther(builder, loc, blockTy.getElementType(),
+                       blockTy.getShape(), paddingOption);
+}
+
+Type getI32TypeLike(OpBuilder &builder, Type ty) {
+  if (auto shapedTy = dyn_cast<ShapedType>(ty))
+    return shapedTy.clone(builder.getI32Type());
+  return builder.getI32Type();
+}
+
+Value getI32ConstLike(OpBuilder &builder, Location loc, Type likeType,
+                      int32_t value) {
+  auto i32Ty = getI32TypeLike(builder, likeType);
+  if (auto shapedTy = dyn_cast<ShapedType>(i32Ty)) {
+    auto attr =
+        DenseElementsAttr::get(shapedTy, builder.getI32IntegerAttr(value));
+    return arith::ConstantOp::create(builder, loc, shapedTy, attr);
+  }
+  return arith::ConstantOp::create(builder, loc, i32Ty,
+                                   builder.getI32IntegerAttr(value));
+}
+
+Value roundF32ToTF32(OpBuilder &builder, Location loc, Value value) {
+  auto valueTy = value.getType();
+  auto i32Ty = getI32TypeLike(builder, valueTy);
+  auto bits = triton::BitcastOp::create(builder, loc, i32Ty, value);
+
+  auto expMask = getI32ConstLike(builder, loc, i32Ty, 0x7F800000);
+  auto exp = arith::AndIOp::create(builder, loc, bits, expMask);
+  auto isSpecial = arith::CmpIOp::create(builder, loc, arith::CmpIPredicate::eq,
+                                         exp, expMask);
+
+  auto shift = getI32ConstLike(builder, loc, i32Ty, 13);
+  auto lsb = arith::AndIOp::create(
+      builder, loc, arith::ShRUIOp::create(builder, loc, bits, shift),
+      getI32ConstLike(builder, loc, i32Ty, 1));
+  auto roundBias = arith::AddIOp::create(
+      builder, loc, lsb, getI32ConstLike(builder, loc, i32Ty, 0x00000FFF));
+  auto rounded = arith::AndIOp::create(
+      builder, loc, arith::AddIOp::create(builder, loc, bits, roundBias),
+      getI32ConstLike(builder, loc, i32Ty, 0xFFFFE000));
+  auto outBits =
+      arith::SelectOp::create(builder, loc, isSpecial, bits, rounded);
+  return triton::BitcastOp::create(builder, loc, valueTy, outBits);
+}
+
+SmallVector<Value> castToI64(OpBuilder &builder, ValueRange values) {
+  auto i64Type = builder.getI64Type();
+  return llvm::map_to_vector(values, [&](Value v) {
+    return builder.createOrFold<arith::ExtSIOp>(v.getLoc(), i64Type, v);
+  });
+}
+
+struct RewriteMakeTensorDesc : OpConversionPattern<triton::MakeTensorDescOp> {
+  using OpConversionPattern<triton::MakeTensorDescOp>::OpConversionPattern;
+
+  llvm::LogicalResult
+  matchAndRewrite(triton::MakeTensorDescOp op, OpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const override {
+
+    SmallVector<Value> ptrShapeStridesPaddingOption;
+    llvm::append_values(ptrShapeStridesPaddingOption, adaptor.getBase());
+    llvm::append_range(ptrShapeStridesPaddingOption,
+                       castToI64(rewriter, adaptor.getShape()));
+    llvm::append_range(ptrShapeStridesPaddingOption, adaptor.getStrides());
+    auto paddingOption = arith::ConstantOp::create(
+        rewriter, op.getLoc(), rewriter.getI1Type(),
+        rewriter.getBoolAttr(adaptor.getPadding() ==
+                             triton::PaddingOption::PAD_NAN));
+    llvm::append_values(ptrShapeStridesPaddingOption, paddingOption);
+    auto roundF32ToTF32 =
+        arith::ConstantOp::create(rewriter, op.getLoc(), rewriter.getI1Type(),
+                                  rewriter.getBoolAttr(false));
+    llvm::append_values(ptrShapeStridesPaddingOption, roundF32ToTF32);
+    rewriter.replaceOpWithMultiple(op, {ptrShapeStridesPaddingOption});
+    return success();
+  }
+};
+
+struct RewriteLoadPattern : OpConversionPattern<triton::DescriptorLoadOp> {
+  using OpConversionPattern<triton::DescriptorLoadOp>::OpConversionPattern;
+
+  llvm::LogicalResult
+  matchAndRewrite(triton::DescriptorLoadOp op, OneToNOpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const override {
+
+    auto loc = op.getLoc();
+    const auto blockShape = op.getDesc().getType().getBlockType().getShape();
+    auto descTy = op.getDesc().getType();
+    auto desc = unpackDescriptor(descTy, adaptor.getDesc());
+    auto offsets = castToI64(rewriter, op.getIndices());
+    auto other = generateOther(rewriter, loc, descTy, desc.paddingOption);
+    auto newLoad = triton::LoadOp::create(
+        rewriter, loc, generatePtr(rewriter, loc, blockShape, desc, offsets),
+        generateMask(rewriter, loc, blockShape, desc, offsets), other,
+        triton::CacheModifier::NONE, triton::EvictionPolicy::NORMAL, false);
+    newLoad->setAttrs(filterSegmentSizes(op->getAttrs()));
+
+    Value result = newLoad.getResult();
+    if (descTy.getBlockType().getElementType().isF32()) {
+
+      auto ifOp = scf::IfOp::create(rewriter, loc, result.getType(),
+                                    desc.roundF32ToTF32, /*withElse=*/true);
+      OpBuilder::InsertionGuard guard(rewriter);
+      rewriter.setInsertionPointToStart(ifOp.thenBlock());
+      auto rounded = roundF32ToTF32(rewriter, loc, result);
+      scf::YieldOp::create(rewriter, loc, rounded);
+
+      rewriter.setInsertionPointToStart(ifOp.elseBlock());
+      scf::YieldOp::create(rewriter, loc, result);
+      result = ifOp.getResult(0);
+    }
+
+    rewriter.replaceOp(op, result);
+    return llvm::success();
+  }
+};
+
+struct RewriteStorePattern : OpConversionPattern<triton::DescriptorStoreOp> {
+  using OpConversionPattern<triton::DescriptorStoreOp>::OpConversionPattern;
+
+  llvm::LogicalResult
+  matchAndRewrite(triton::DescriptorStoreOp op, OneToNOpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const override {
+
+    auto loc = op.getLoc();
+    auto descTy = op.getDesc().getType();
+    const auto blockShape = descTy.getBlockType().getShape();
+    auto desc = unpackDescriptor(descTy, adaptor.getDesc());
+    auto offsets = castToI64(rewriter, op.getIndices());
+
+    auto newStore = rewriter.replaceOpWithNewOp<triton::StoreOp>(
+        op, generatePtr(rewriter, loc, blockShape, desc, offsets), op.getSrc(),
+        generateMask(rewriter, loc, blockShape, desc, offsets),
+        triton::CacheModifier::NONE, triton::EvictionPolicy::NORMAL);
+    newStore->setAttrs(filterSegmentSizes(op->getAttrs()));
+
+    return llvm::success();
+  }
+};
+
+std::pair<Value, Value>
+generateGatherScatterPtrMask(OpBuilder &builder, Location loc,
+                             ArrayRef<int64_t> blockShape, Descriptor &desc,
+                             Value xOffsets, Value yOffset) {
+  Value xOffsetRange =
+      expandOffsets(builder, loc, blockShape, xOffsets, /*dim=*/0);
+  yOffset = castToI64(builder, {yOffset})[0];
+  auto xOffsetI64Ty = RankedTensorType::get(
+      cast<RankedTensorType>(xOffsetRange.getType()).getShape(),
+      yOffset.getType());
+  xOffsetRange =
+      arith::ExtSIOp::create(builder, loc, xOffsetI64Ty, xOffsetRange);
+  auto yOffsetRange =
+      getExpandedOffsetWithRange(builder, loc, blockShape, yOffset, /*dim=*/1);
+  auto ptr = generatePtrFromOffsetRanges(builder, loc, blockShape, desc,
+                                         {xOffsetRange, yOffsetRange});
+  auto mask = generateMaskFromOffsetRanges(builder, loc, blockShape, desc,
+                                           {xOffsetRange, yOffsetRange});
+  return {ptr, mask};
+}
+
+struct RewriteGatherPattern : OpConversionPattern<triton::DescriptorGatherOp> {
+  using OpConversionPattern<triton::DescriptorGatherOp>::OpConversionPattern;
+
+  llvm::LogicalResult
+  matchAndRewrite(triton::DescriptorGatherOp op, OneToNOpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const override {
+    auto loc = op.getLoc();
+    auto descTy = op.getDesc().getType();
+    const auto blockShape = op.getResult().getType().getShape();
+    auto desc = unpackDescriptor(descTy, adaptor.getDesc());
+    auto [ptr, mask] = generateGatherScatterPtrMask(
+        rewriter, loc, blockShape, desc, op.getXOffsets(), op.getYOffset());
+    auto other = generateOther(rewriter, loc,
+                               descTy.getSignlessBlockType().getElementType(),
+                               blockShape, desc.paddingOption);
+    auto newLoad = triton::LoadOp::create(
+        rewriter, loc, ptr, mask, other, triton::CacheModifier::NONE,
+        triton::EvictionPolicy::NORMAL, false);
+    newLoad->setAttrs(filterSegmentSizes(op->getAttrs()));
+
+    Value result = newLoad.getResult();
+    if (descTy.getSignlessBlockType().getElementType().isF32()) {
+      auto rounded = roundF32ToTF32(rewriter, loc, result);
+      result = arith::SelectOp::create(rewriter, loc, desc.roundF32ToTF32,
+                                       rounded, result);
+    }
+
+    rewriter.replaceOp(op, result);
+    return llvm::success();
+  }
+};
+
+struct RewriteScatterPattern
+    : OpConversionPattern<triton::DescriptorScatterOp> {
+  using OpConversionPattern<triton::DescriptorScatterOp>::OpConversionPattern;
+
+  llvm::LogicalResult
+  matchAndRewrite(triton::DescriptorScatterOp op, OneToNOpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const override {
+    auto loc = op.getLoc();
+    auto descTy = op.getDesc().getType();
+    const auto blockShape = op.getSrc().getType().getShape();
+    auto desc = unpackDescriptor(descTy, adaptor.getDesc());
+    auto [ptr, mask] = generateGatherScatterPtrMask(
+        rewriter, loc, blockShape, desc, op.getXOffsets(), op.getYOffset());
+    auto newStore = rewriter.replaceOpWithNewOp<triton::StoreOp>(
+        op, ptr, op.getSrc(), mask, triton::CacheModifier::NONE,
+        triton::EvictionPolicy::NORMAL);
+    newStore->setAttrs(filterSegmentSizes(op->getAttrs()));
+
+    return llvm::success();
+  }
+};
+
+std::optional<RMWOp> translateReduceKind(DescriptorReduceKind kind,
+                                         TensorDescType ty) {
+  auto scalarTy = ty.getBlockType().getElementType();
+  switch (kind) {
+  case DescriptorReduceKind::ADD:
+    return scalarTy.isInteger() ? RMWOp::ADD : RMWOp::FADD;
+  case DescriptorReduceKind::MIN:
+    if (scalarTy.isUnsignedInteger()) {
+      return RMWOp::UMIN;
+    } else if (scalarTy.isSignedInteger()) {
+      return RMWOp::MIN;
+    }
+    return {};
+  case DescriptorReduceKind::MAX:
+    if (scalarTy.isUnsignedInteger()) {
+      return RMWOp::UMAX;
+    } else if (scalarTy.isSignedInteger()) {
+      return RMWOp::MAX;
+    }
+    return {};
+  case DescriptorReduceKind::AND:
+    return RMWOp::AND;
+  case DescriptorReduceKind::OR:
+    return RMWOp::OR;
+  case DescriptorReduceKind::XOR:
+    return RMWOp::XOR;
+  default:
+    break;
+  }
+  return {};
+}
+
+struct RewriteReducePattern : OpConversionPattern<triton::DescriptorReduceOp> {
+  using OpConversionPattern<triton::DescriptorReduceOp>::OpConversionPattern;
+
+  llvm::LogicalResult
+  matchAndRewrite(triton::DescriptorReduceOp op, OneToNOpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const override {
+    auto loc = op.getLoc();
+    auto descTy = op.getDesc().getType();
+    const auto blockShape = descTy.getBlockType().getShape();
+    auto desc = unpackDescriptor(descTy, adaptor.getDesc());
+    auto offsets = castToI64(rewriter, op.getIndices());
+    auto rmwOp = translateReduceKind(op.getKind(), descTy);
+    if (!rmwOp) {
+      std::string msgstring;
+      llvm::raw_string_ostream msg(msgstring);
+      msg << "Cannot fallback on descriptor atomic op, unsupported for type "
+          << descTy.getBlockType().getElementType();
+      return op->emitError(msgstring);
+    }
+
+    triton::AtomicRMWOp::create(
+        rewriter, loc, descTy.getSignlessBlockType(), *rmwOp,
+        generatePtr(rewriter, loc, blockShape, desc, offsets), op.getSrc(),
+        generateMask(rewriter, loc, blockShape, desc, offsets),
+        MemSemantic::RELEASE, MemSyncScope::GPU);
+    op.erase();
+    return success();
+  }
+};
+} // namespace
+
+bool hasATensorDescriptorType(TypeRange types) {
+  return llvm::any_of(
+      types, [](Type t) { return llvm::isa<triton::TensorDescType>(t); });
+}
+
+LogicalResult convertTensorDescTypeToPointer(triton::TensorDescType t,
+                                             llvm::SmallVectorImpl<Type> &out) {
+  // We convert a tensor descriptor into an pointer, and a shape and stride
+  // for each dimension, and padding option. i.e., we create 1+2*rank+1
+  // values. Note that tensor descriptors may be signed/unsigned integers
+  // whereas pointers should always be signless.
+  auto tensorType = t.getSignlessBlockType();
+  out.push_back(triton::getPointerType(tensorType.getElementType()));
+  out.insert(out.end(), 2 * tensorType.getRank(),
+             IntegerType::get(t.getContext(), 64));
+  out.push_back(IntegerType::get(t.getContext(), 1));
+  out.push_back(IntegerType::get(t.getContext(), 1));
+  return success();
+}
+
+TypeConverter createDescTypeConverter() {
+  TypeConverter converter;
+
+  converter.addConversion([](Type t) {
+    // Most types don't require any conversion
+    return t;
+  });
+  converter.addConversion(convertTensorDescTypeToPointer);
+
+  return converter;
+}
+
+FuncArgRenamer createFunArgRenamer() {
+  FuncArgRenamer renamer(".");
+  renamer.addRenamer([](triton::TensorDescType type,
+                        llvm::SmallVectorImpl<std::string> &out_suffix) {
+    auto tensorType = type.getSignlessBlockType();
+    int dims = tensorType.getRank();
+    out_suffix.push_back("");
+    for (int i = 0; i < dims; i++) {
+      out_suffix.push_back("shape." + std::to_string(i));
+    }
+    for (int i = 0; i < dims; i++) {
+      out_suffix.push_back("stride." + std::to_string(i));
+    }
+    out_suffix.push_back("padding");
+    out_suffix.push_back("roundF32ToTF32");
+    return success();
+  });
+
+  return renamer;
+}
+
+void populateMakeTensorDescriptorPattern(RewritePatternSet &patterns,
+                                         TypeConverter &converter) {
+  patterns.add<RewriteMakeTensorDesc>(converter, patterns.getContext());
+}
+
+void populateLoadTensorDescriptorPattern(RewritePatternSet &patterns,
+                                         TypeConverter &converter) {
+  patterns.add<RewriteLoadPattern>(converter, patterns.getContext());
+}
+
+void populateStoreTensorDescriptorPattern(RewritePatternSet &patterns,
+                                          TypeConverter &converter) {
+  patterns.add<RewriteStorePattern>(converter, patterns.getContext());
+}
+
+void populateGatherDescriptorPattern(RewritePatternSet &patterns,
+                                     TypeConverter &converter) {
+  patterns.add<RewriteGatherPattern>(converter, patterns.getContext());
+}
+
+void populateScatterTensorDescriptorPattern(RewritePatternSet &patterns,
+                                            TypeConverter &converter) {
+  patterns.add<RewriteScatterPattern>(converter, patterns.getContext());
+}
+
+void populateReduceTensorDescriptorPattern(RewritePatternSet &patterns,
+                                           TypeConverter &converter) {
+  patterns.add<RewriteReducePattern>(converter, patterns.getContext());
+}
+} // namespace mlir::triton
